+++ CHANGELOG.md
@@ -4,8 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
-- Added unit tests to TrainData helper methods
+## [0.3.6] - 2025-01-22
+- Added unit tests to TrainData helper methods.
+- Added option to enable/disable progress in PyTorch model training loop
+- Added abstract base class for TrainCallbacks
+- Added abstract base class and simple printer for StepCallbacks
+- Added a callback to be called after each batch to the training loop
 
 ## [0.3.5] - 2025-01-06
 - Exponential smoothing per-batch loss also for validation errors